requirements.txt
streamlit==1.28.1
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
jieba==0.42.1
matplotlib==3.7.2
reportlab==4.0.4
plotly==5.17.0
pandas==2.1.1
streamlit-cropper==0.2.1
streamlit-image-coordinates==0.1.6